updated script to support macOS

+++ jinstall.sh
@@ -28,28 +28,40 @@ WD=~/.tmp_update #Working Directory
 function jtinstall {
         wget http://www.newosxbook.com/tools/jtool.tar
         tar xvf jtool.tar
-        rm jtool
-        mv jtool.ELF64 jtool
-		sudo cp jtool.1 /usr/local/share/man/man1
-        sudo cp jtool $INSTALL_DIR
-        echo -e "\033[1;32mInstalled jtool to $INSTALL_DIR"
+        if [[ $(uname) == 'Darwin' ]]; then
+			cp jtool /usr/local/bin
+		else
+			rm jtool
+			mv jtool.ELF64 jtool
+			sudo cp jtool.1 /usr/local/share/man/man1
+			sudo cp jtool $INSTALL_DIR
+		fi
+		echo -e "\033[1;32mInstalled jtool to $INSTALL_DIR"
 }
 
 function jokerinstall {
         wget http://www.newosxbook.com/tools/joker.tar
         tar xvf joker.tar
-        mv joker.ELF64 joker
-        sudo cp joker $INSTALL_DIR
-        echo -e "\033[1;32mInstalled joker to $INSTALL_DIR"
+        if [[ $(uname) == 'Darwin' ]]; then
+			cp joker.universal /usr/local/bin/joker
+		else
+			mv joker.ELF64 joker
+			sudo cp joker $INSTALL_DIR
+        fi
+		echo -e "\033[1;32mInstalled joker to $INSTALL_DIR"
 }
 
 function disarm_install {
         wget http://newosxbook.com/tools/disarm.tar
         tar xvf disarm.tar
-        rm disarm
-        mv disarm.ELF64 disarm
-        sudo cp disarm $INSTALL_DIR
-        echo -e "\033[1;32mInstalled disarm to $INSTALL_DIR"
+        if [[ $(uname) == 'Darwin' ]]; then
+			cp disarm /usr/local/bin/
+		else
+			rm disarm
+			mv disarm.ELF64 disarm
+			sudo cp disarm $INSTALL_DIR
+        fi
+		echo -e "\033[1;32mInstalled disarm to $INSTALL_DIR"
 }
 
 function otastuff {
@@ -66,13 +78,15 @@ function otastuff {
 }
 
 function main {
-        if [[ $(uname) != 'Linux' && $(arch) != 'x86_64' ]]; then
-                echo "This script is only for Linux 64 bits"
-        exit 1
-        fi
+        if [[ $(uname) == 'Linux' && $(arch) == 'x86_64' || $(uname) == 'Darwin' ]]; then
+			echo "starting script..."
+		else
+			echo "This script is only for Linux 64 bits or macOS"
+			exit 1
+		fi
         if [[ -e $WD ]]; then
                 rm -rf $WD
-        fi
+		fi
         mkdir $WD
         cd $WD
         jtinstall
